Use maven to build and release
