+++ .github/workflows/release.yml
@@ -9,7 +9,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v2
+        uses: actions/checkout@v3
         with:
           fetch-depth: 5
           persist-credentials: false
@@ -18,46 +18,47 @@ jobs:
         uses: actions/setup-java@v1
         with:
           java-version: 11
+          distribution: 'temurin'
+          server-id: ossrh
+          server-username: MAVEN_USERNAME
+          server-password: MAVEN_CENTRAL_TOKEN
+          gpg-private-key: ${{ secrets.GPG_PRIVATE_KEY }}
+          gpg-passphrase: MAVEN_GPG_PASSPHRASE
 
       - name: Configure GIT
         run: |
           git config --global user.name "Gluon Bot"
           git config --global user.email "githubbot@gluonhq.com"
 
-      - name: Import GPG keys
-        run: |
-          gpg --import --batch <(echo "$GPG_PRIVATE_KEY") &> /dev/null
-          # Gradle doesn't support GPG 2.1 and later: https://github.com/gradle/gradle/issues/888
-          gpg --export-secret-keys --pinentry-mode loopback --passphrase="$GPG_PASSPHRASE" > ~/.gnupg/secring.gpg
-          rm -rf /tmp/secret
-        env:
-          GPG_PRIVATE_KEY: ${{ secrets.GPG_PRIVATE_KEY }}
-          GPG_PASSPHRASE: ${{ secrets.GPG_PASSPHRASE }}
+      - name: Grant execute permission for maven wrapper
+        run: chmod +x mvnw
 
-      - name: Release to Maven Central
+      - name: Publish to Maven Central
         id: deploy
         run: |
-          ./gradlew publishToSonatype closeAndReleaseSonatypeStagingRepository -i \
-          -PsonatypeUsername=$SONATYPE_USERNAME -PsonatypePassword=$SONATYPE_PASSWORD \
-          -Psigning.keyId=$GPG_KEYNAME -Psigning.password=$GPG_PASSPHRASE -Psigning.secretKeyRingFile=$HOME/.gnupg/secring.gpg
+          ./mvnw -DskipTests=true -B -U -ntp -Prelease deploy -pl '!samples'
           echo ::set-output name=exit_code::$?
         env:
-          SONATYPE_USERNAME: ${{ secrets.SONATYPE_USERNAME }}
-          SONATYPE_PASSWORD: ${{ secrets.SONATYPE_PASSWORD }}
-          GPG_PASSPHRASE: ${{ secrets.GPG_PASSPHRASE }}
-          GPG_KEYNAME: ${{ secrets.GPG_KEYNAME }}
+          MAVEN_USERNAME: ${{ secrets.SONATYPE_USERNAME }}
+          MAVEN_CENTRAL_TOKEN: ${{ secrets.SONATYPE_PASSWORD }}
+          MAVEN_GPG_PASSPHRASE: ${{ secrets.GPG_PASSPHRASE }}
+
+      - name: Create GitHub release
+        if: steps.deploy.outputs.exit_code == 0
+        uses: softprops/action-gh-release@v1
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+          generate_release_notes: true
 
       - name: Commit next development version
         if: steps.deploy.outputs.exit_code == 0
         run: |
           TAG=${GITHUB_REF/refs\/tags\//}
-          if [[ $TAG == *"ea+"* ]]; then
-            newVersion=${TAG%+*}+$((${TAG##*+} + 1)) # Update ea version by 1
-          else
-            newVersion=${TAG%.*}.$((${TAG##*.} + 1)) # Update version by 1
-          fi
-          sed -i -z "0,/version = $TAG/s//version = $newVersion-SNAPSHOT/" gradle.properties
-          git commit gradle.properties -m "Prepare development of $newVersion"
-          git push https://gluon-bot:$PAT@github.com/$GITHUB_REPOSITORY HEAD:master
+          NEW_PROJECT_VERSION=${TAG%.*}.$((${TAG##*.} + 1))
+          echo "Update project version to next snapshot version"
+          ./mvnw versions:set -DnewVersion=$NEW_PROJECT_VERSION-SNAPSHOT -DgenerateBackupPoms=false
+          git commit pom.xml -m "Prepare development of $NEW_PROJECT_VERSION"
+          git push https://gluon-bot:$PAT@github.com/$GITHUB_REPOSITORY HEAD:main
+        shell: bash
         env:
-          PAT: ${{ secrets.PAT }}
+          PAT: ${{ secrets.GITHUB_TOKEN }}
