--- gradle.properties
@@ -1,2 +1,0 @@
-group   = com.gluonhq
-version = 2.0.0-ea+7-SNAPSHOT
