--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Tue Nov 15 21:17:36 CET 2016
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-7.0.1-bin.zip
