--- mavenPublish.gradle
@@ -1,93 +1,0 @@
-/*
- * Copyright (c) 2016, 2022, Gluon
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL GLUON BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-apply plugin: 'maven-publish'
-apply plugin: 'signing'
-
-java {
-    withJavadocJar()
-    withSourcesJar()
-}
-
-publishing {
-
-    publications {
-        mavenJava(MavenPublication) {
-
-            from components.java
-
-            pom {
-                name = 'Gluon Maps'
-                description = "Gluon Maps"
-                url = 'https://gluonhq.com/labs/maps'
-                scm {
-                    connection = 'scm:git:https://github.com/gluonhq/maps'
-                    developerConnection = 'scm:git:https://github.com/gluonhq/maps'
-                    url = 'https://github.com/gluonhq/maps'
-                }
-                developers {
-                    developer {
-                        name = 'Johan Vos'
-                        email = 'johan.vos@gluonhq.com'
-                        organization = 'Gluon'
-                        organizationUrl = 'http://gluonhq.com'
-                    }
-                }
-                licenses {
-                    license {
-                        name = 'GNU General Public License (GPL)'
-                        url = 'http://www.gnu.org/licenses/gpl.txt'
-                    }
-                }
-            }
-
-            pom.withXml {
-                def pomNode = asNode()
-                pomNode.dependencies.'*'.findAll() {
-                    it.groupId.text() == 'org.openjfx'
-                }.each {
-                    it.remove(it.classifier)
-                }
-            }
-        }
-    }
-}
-
-signing {
-    required = { gradle.taskGraph.hasTask('publishing') && !version.endsWith("SNAPSHOT") }
-    sign publishing.publications.mavenJava
-}
-
-nexusPublishing {
-    repositories {
-        sonatype {
-            username = project.hasProperty('sonatypeUsername') ? project.property('sonatypeUsername') : ''
-            password = project.hasProperty('sonatypePassword') ? project.property('sonatypePassword') : ''
-        }
-    }
-}
