+++ samples/src/main/java/com/gluonhq/maps/samples/PoiLayer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2021, Gluon
+ * Copyright (c) 2016, 2023, Gluon
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -25,18 +25,44 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package com.gluonhq.maps.samples;
 
-plugins {
-    id 'com.github.ben-manes.versions' version '0.39.0'
-    id 'com.github.hierynomus.license' version '0.16.1'
-    id 'net.nemerosa.versioning' version '2.14.0'
-}
+import com.gluonhq.maps.MapLayer;
+import com.gluonhq.maps.MapPoint;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.geometry.Point2D;
+import javafx.scene.Node;
+import javafx.util.Pair;
 
-apply from: rootProject.file('mavenPublish.gradle')
-apply from: rootProject.file('gradle/publishing.gradle')
-apply from: rootProject.file('gradle/code-quality.gradle')
+/**
+ *
+ * A layer that allows to visualise points of interest.
+ */
+public class PoiLayer extends MapLayer {
 
-dependencies {
-    api "com.gluonhq.attach:storage:$attachVersion"
-    api "com.gluonhq.attach:util:$attachVersion"
+
+    private final ObservableList<Pair<MapPoint, Node>> points = FXCollections.observableArrayList();
+    
+    public PoiLayer() {
+    }
+
+    public void addPoint(MapPoint p, Node icon) {
+        points.add(new Pair<>(p, icon));
+        this.getChildren().add(icon);
+        this.markDirty();
+    }
+
+    @Override
+    protected void layoutLayer() {
+        for (Pair<MapPoint, Node> candidate : points) {
+            MapPoint point = candidate.getKey();
+            Node icon = candidate.getValue();
+            Point2D mapPoint = getMapPoint(point.getLatitude(), point.getLongitude());
+            icon.setVisible(true);
+            icon.setTranslateX(mapPoint.getX());
+            icon.setTranslateY(mapPoint.getY());
+        }
+    }
+
 }
