+++ samples/src/main/resources/logging.properties
